ACTIG: score-period mode, extended fragmentation metrics, QC improvements
- Add score-period mode: scores and writes SP/WP period-window annotations
  only (no metric output); uses coarser defaults suited for defining
  sleep/wake period windows in a subsequent ACTIG score run
- Add sleep-period/wake-period annotation support: fragmentation metrics
  now restricted to SP/WP windows; new tp-lambda, thresh-smooth params
- New fragmentation metrics: TP_SW/TP_WS transition probabilities (Bayesian
  MLE), TRANS_ENT (Shannon conditional entropy), RL_ENT (run-length entropy)
- Rename gap annotation ACTIG_G -> G (score) / GP (score-period)
- New QC params: qc-exclude-allsleep/allwake, qc-allsleep-th/allwake-th,
  qc-extreme-out; updated qc-exclude-out default label
- Add verbose and investigate modes for diagnostics and parameter tuning
- DAYS: add bin (clock-time histogram bins), window (named clock windows),
  days=F to suppress day annotations
- TABULATE: add histogram binning via bin=N with min/max bounds
- LOG: add offset= param for fixed additive constant; improve eps handling
- hypno: add SFI1 (transitions into W or N1) fragmentation metric
- functions.cpp: fix Fs type from int to double

+++ timeline/functions.cpp
@@ -2471,7 +2471,7 @@ void timeline_t::signal_means_by_annot( const param_t & param )
 
   if ( ns == 0 ) return;
   
-  const int Fs = edf->header.sampling_freq( signals(0) );
+  const double Fs = edf->header.sampling_freq( signals(0) );
   for (int s=1; s<ns; s++)
     if ( edf->header.sampling_freq( signals(s) ) != Fs )
       Helper::halt( "signals must have similar sampling rates" );
